Simplify and consolidate fixtures
