+++ jwst/saturation/tests/test_saturation.py
@@ -2,15 +2,15 @@
 
 import numpy as np
 import pytest
-from stdatamodels.jwst.datamodels import RampModel, SaturationModel, SuperBiasModel, dqflags
+from stdatamodels.jwst.datamodels import dqflags
 
-from jwst.dq_init.tests.helpers import make_superstripe_model
 from jwst.lib import reffile_utils
 from jwst.saturation import SaturationStep
 from jwst.saturation.saturation import flag_saturation, irs2_flag_saturation
+from jwst.saturation.tests import helpers
 
 
-def test_basic_saturation_flagging(setup_nrc_cube):
+def test_basic_saturation_flagging():
     """Check that the saturation flag is set when a pixel value is above the
     threshold given by the reference file."""
 
@@ -20,7 +20,7 @@ def test_basic_saturation_flagging(setup_nrc_cube):
     ncols = 20
     satvalue = 60000
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values up to the saturation limit
     data.data[0, 0, 5, 5] = 0
@@ -40,8 +40,8 @@ def test_basic_saturation_flagging(setup_nrc_cube):
     assert np.all(output.groupdq[0, satindex:, 4:6, 4:6] == dqflags.group["SATURATED"])
 
 
-def test_nirspec_irs2_saturation_flagging(setup_nrs_irs2_cube):
-    data, satmap, _ = setup_nrs_irs2_cube()
+def test_nirspec_irs2_saturation_flagging():
+    data, satmap, _ = helpers.setup_nrs_irs2_cube()
 
     pixx, pixy = 1000, 1000
     data.data[0, 3, pixx, pixy] = 65000  # Signal exceeds saturation limit of 60000
@@ -57,11 +57,11 @@ def test_nirspec_irs2_saturation_flagging(setup_nrs_irs2_cube):
     )
 
 
-def test_nirspec_irs2_readpatt(setup_nrs_irs2_cube):
+def test_nirspec_irs2_readpatt():
     """
     Tests that the readpatt framework (saturation in grouped data) is working for IRS2 processing.
     """
-    data, satmap, _ = setup_nrs_irs2_cube()
+    data, satmap, _ = helpers.setup_nrs_irs2_cube()
 
     pixx, pixy = 1000, 1000
     data.data[0, 0, pixx, pixy] = 500  # Low signal
@@ -77,11 +77,11 @@ def test_nirspec_irs2_readpatt(setup_nrs_irs2_cube):
     assert np.all(output.groupdq[0, 1, pixx, pixy] == dqflags.group["SATURATED"])
 
 
-def test_nirspec_irs2_group2_sat_bias(setup_nrs_irs2_cube):
+def test_nirspec_irs2_group2_sat_bias():
     """
     Test group 2 saturation check with bias in grouped IRS2 data.
     """
-    data, satmap, bias = setup_nrs_irs2_cube()
+    data, satmap, bias = helpers.setup_nrs_irs2_cube()
 
     pixx, pixy = 1000, 1000
     data.data[0, 0, pixx, pixy] = 18000  # 15000 bias + 1000 counts per frame
@@ -99,12 +99,12 @@ def test_nirspec_irs2_group2_sat_bias(setup_nrs_irs2_cube):
     assert np.all(output.groupdq[0, 1, pixx, pixy] == dqflags.group["SATURATED"])
 
 
-def test_nirspec_nrs_readpatt(setup_nrs_nrs_cube):
+def test_nirspec_nrs_readpatt():
     """
     Tests that the readpatt framework (saturation in grouped data) is working for regular grouped
     data (code follows all instruments except NIRSpec IRS2)
     """
-    data, satmap = setup_nrs_nrs_cube()
+    data, satmap = helpers.setup_nrs_nrs_cube()
 
     pixx, pixy = 1000, 1000
     data.data[0, 0, pixx, pixy] = 500  # Low signal
@@ -120,14 +120,14 @@ def test_nirspec_nrs_readpatt(setup_nrs_nrs_cube):
     assert np.all(output.groupdq[0, 1, pixx, pixy] == dqflags.group["SATURATED"])
 
 
-def test_irs2_zero_frame(setup_nrs_irs2_cube):
+def test_irs2_zero_frame():
     """
     Tests IRS2 ZEROFRAME processing.
 
     This ensures ZEROFRAME data that is outside saturation and AD floor
     boundaries get zeroed out.
     """
-    ramp, sat, _ = setup_nrs_irs2_cube()
+    ramp, sat, _ = helpers.setup_nrs_irs2_cube()
 
     # Setup ZEROFRAME
     nints, ngroups, nrows, ncols = ramp.data.shape
@@ -148,7 +148,7 @@ def test_irs2_zero_frame(setup_nrs_irs2_cube):
     np.testing.assert_allclose(output.zeroframe, check_zframe, tol)
 
 
-def test_ad_floor_flagging(setup_nrc_cube):
+def test_ad_floor_flagging():
     """Check that the ad_floor flag is set when a pixel value is zero or
     negative."""
 
@@ -158,7 +158,7 @@ def test_ad_floor_flagging(setup_nrc_cube):
     ncols = 20
     satvalue = 60000
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values up to the saturation limit
     data.data[0, 0, 5, 5] = 0  # Signal at bottom rail - low saturation
@@ -182,7 +182,7 @@ def test_ad_floor_flagging(setup_nrc_cube):
     )
 
 
-def test_ad_floor_and_saturation_flagging(setup_nrc_cube):
+def test_ad_floor_and_saturation_flagging():
     """Check that the ad_floor flag is set when a pixel value is zero or
     negative and the saturation flag when the pixel is above the saturation threshold."""
 
@@ -192,7 +192,7 @@ def test_ad_floor_and_saturation_flagging(setup_nrc_cube):
     ncols = 20
     satvalue = 60000
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values up to the saturation limit
     data.data[0, 0, 5, 5] = 0  # Signal at bottom rail - low saturation
@@ -224,7 +224,7 @@ def test_ad_floor_and_saturation_flagging(setup_nrc_cube):
     )
 
 
-def test_signal_fluctuation_flagging(setup_nrc_cube):
+def test_signal_fluctuation_flagging():
     """Check that once a pixel is flagged as saturated in a group, all
     subsequent groups should also be flagged as saturated, even if the
     signal value drops back below saturation."""
@@ -235,7 +235,7 @@ def test_signal_fluctuation_flagging(setup_nrc_cube):
     ncols = 20
     satvalue = 60000
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values up to the saturation limit
     data.data[0, 0, 5, 5] = 10
@@ -258,7 +258,7 @@ def test_signal_fluctuation_flagging(setup_nrc_cube):
     )
 
 
-def test_all_groups_saturated(setup_nrc_cube):
+def test_all_groups_saturated():
     """Check case where all groups are saturated."""
 
     # Create inputs, data, and saturation maps
@@ -267,7 +267,7 @@ def test_all_groups_saturated(setup_nrc_cube):
     ncols = 20
     satvalue = 60000
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values at or above saturation limit
     data.data[0, 0, 5, 5] = 60000
@@ -289,7 +289,7 @@ def test_all_groups_saturated(setup_nrc_cube):
     )
 
 
-def test_subarray_extraction(setup_miri_cube):
+def test_subarray_extraction():
     """Check the step correctly handles subarrays."""
 
     # Create input data
@@ -298,7 +298,7 @@ def test_subarray_extraction(setup_miri_cube):
     nrows = 224
     ncols = 288
 
-    data, satmap = setup_miri_cube(1, 467, ngroups, nrows, ncols)
+    data, satmap = helpers.setup_miri_cube(1, 467, ngroups, nrows, ncols)
 
     # Place DQ flags in DQ array that would be in subarray
     # MASK1550 file has colstart=1, rowstart=467
@@ -321,7 +321,7 @@ def test_subarray_extraction(setup_miri_cube):
     assert output.pixeldq[84, 100] == dqflags.pixel["NO_SAT_CHECK"] + dqflags.pixel["NONLINEAR"]
 
 
-def test_dq_propagation(setup_nrc_cube):
+def test_dq_propagation():
     """Check PIXELDQ propagation."""
 
     # Create inputs, data, and saturation maps
@@ -331,7 +331,7 @@ def test_dq_propagation(setup_nrc_cube):
     dqval1 = 5
     dqval2 = 10
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add DQ values to the data and reference file
     data.pixeldq[5, 5] = dqval1
@@ -344,7 +344,7 @@ def test_dq_propagation(setup_nrc_cube):
     assert output.pixeldq[5, 5] == dqval1 + dqval2
 
 
-def test_no_sat_check(setup_nrc_cube):
+def test_no_sat_check():
     """Check that pixels flagged with NO_SAT_CHECK in the reference file get
     added to the DQ mask and are not flagged as saturated."""
 
@@ -354,7 +354,7 @@ def test_no_sat_check(setup_nrc_cube):
     ncols = 20
     satvalue = 60000
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values up to the saturation limit
     data.data[0, 0, 5, 5] = 10
@@ -379,7 +379,7 @@ def test_no_sat_check(setup_nrc_cube):
     assert output.pixeldq[5, 5] == (dqflags.pixel["NO_SAT_CHECK"] + dqflags.pixel["RC"])
 
 
-def test_nans_in_mask(setup_nrc_cube):
+def test_nans_in_mask():
     """Check that pixels in the reference files that have value NaN are not
     flagged as saturated in the data and that in the PIXELDQ array the
     pixel is set to NO_SAT_CHECK."""
@@ -389,7 +389,7 @@ def test_nans_in_mask(setup_nrc_cube):
     nrows = 20
     ncols = 20
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # Add ramp values up to the saturation limit
     data.data[0, 0, 5, 5] = 10
@@ -410,7 +410,7 @@ def test_nans_in_mask(setup_nrc_cube):
     assert output.pixeldq[5, 5] == dqflags.pixel["NO_SAT_CHECK"]
 
 
-def test_full_step(setup_nrc_cube):
+def test_full_step():
     """Test full run of the SaturationStep."""
 
     # Create inputs, data, and saturation maps
@@ -418,7 +418,7 @@ def test_full_step(setup_nrc_cube):
     nrows = 20
     ncols = 20
 
-    data, satmap = setup_nrc_cube(ngroups, nrows, ncols)
+    data, satmap = helpers.setup_nrc_cube(ngroups, nrows, ncols)
 
     # set the entire array to a small non-zero value to avoid labeling
     # almost everything as low saturated
@@ -449,11 +449,11 @@ def test_full_step(setup_nrc_cube):
     assert data.meta.cal_step.saturation is None
 
 
-def test_full_step_irs2(setup_nrs_irs2_cube):
+def test_full_step_irs2():
     """Test full run of the SaturationStep for IRS2 data."""
 
     # Create input data, saturation map, and bias
-    data, satmap, bias = setup_nrs_irs2_cube()
+    data, satmap, bias = helpers.setup_nrs_irs2_cube()
 
     pixx, pixy = 1000, 1000
     data.data[0, 0, pixx, pixy] = 18000  # 15000 bias + 1000 counts per frame
@@ -477,8 +477,8 @@ def test_full_step_irs2(setup_nrs_irs2_cube):
     assert data.meta.cal_step.saturation is None
 
 
-def test_skip_missing_reffile(setup_nrc_cube):
-    data, _ = setup_nrc_cube(5, 20, 20)
+def test_skip_missing_reffile():
+    data, _ = helpers.setup_nrc_cube(5, 20, 20)
     output = SaturationStep.call(data, override_saturation="N/A")
 
     # Step is marked skipped
@@ -490,8 +490,8 @@ def test_skip_missing_reffile(setup_nrc_cube):
 
 
 @pytest.mark.parametrize("use_bias", [True, False])
-def test_superstripe(setup_nis_superstripe_cube, use_bias):
-    data, satmap, bias_model = setup_nis_superstripe_cube()
+def test_superstripe(use_bias):
+    data, satmap, bias_model = helpers.setup_nis_superstripe_cube()
 
     # Add ramp values up to the saturation limit
     satvalue = 60000
@@ -531,216 +531,3 @@ def test_superstripe(setup_nis_superstripe_cube, use_bias):
     # Make sure the pixeldq is reset properly. No new flags are expected.
     assert output.pixeldq.shape == (nstripe, *data.data.shape[-2:])
     assert np.all(output.pixeldq == 0)
-
-
-def add_test_refmodel_metadata(refmodel):
-    refmodel.meta.telescope = "JWST"
-    refmodel.meta.description = "filler"
-    refmodel.meta.reftype = "filler"
-    refmodel.meta.author = "Py Test"
-    refmodel.meta.pedigree = "Pytest"
-    refmodel.meta.useafter = "2015-01-01T01:00:00"
-
-
-@pytest.fixture(scope="function")
-def setup_nrc_cube():
-    """Set up fake NIRCam data to test."""
-
-    def _cube(ngroups, nrows, ncols):
-        nints = 1
-
-        data_model = RampModel((nints, ngroups, nrows, ncols))
-        data_model.meta.subarray.xstart = 1
-        data_model.meta.subarray.ystart = 1
-        data_model.meta.subarray.xsize = ncols
-        data_model.meta.subarray.ysize = nrows
-        data_model.meta.exposure.ngroups = ngroups
-        data_model.meta.exposure.nframes = 3
-        data_model.meta.instrument.name = "NIRCAM"
-        data_model.meta.instrument.detector = "NRCA1"
-        data_model.meta.observation.date = "2017-10-01"
-        data_model.meta.observation.time = "00:00:00"
-
-        saturation_model = SaturationModel((2048, 2048))
-        saturation_model.meta.subarray.xstart = 1
-        saturation_model.meta.subarray.ystart = 1
-        saturation_model.meta.subarray.xsize = 2048
-        saturation_model.meta.subarray.ysize = 2048
-        saturation_model.meta.instrument.name = "NIRCAM"
-        add_test_refmodel_metadata(saturation_model)
-
-        return data_model, saturation_model
-
-    return _cube
-
-
-@pytest.fixture(scope="function")
-def setup_miri_cube():
-    """Set up fake MIRI data to test."""
-
-    def _cube(xstart, ystart, ngroups, nrows, ncols):
-        nints = 1
-
-        # create a JWST datamodel for MIRI data
-        data_model = RampModel((nints, ngroups, nrows, ncols))
-        data_model.data += 1
-        data_model.meta.instrument.name = "MIRI"
-        data_model.meta.instrument.detector = "MIRIMAGE"
-        data_model.meta.instrument.filter = "F1500W"
-        data_model.meta.instrument.band = "N/A"
-        data_model.meta.observation.date = "2016-06-01"
-        data_model.meta.observation.time = "00:00:00"
-        data_model.meta.exposure.type = "MIR_IMAGE"
-        data_model.meta.subarray.name = "MASK1550"
-        data_model.meta.subarray.xstart = xstart
-        data_model.meta.subarray.xsize = ncols
-        data_model.meta.subarray.ystart = ystart
-        data_model.meta.subarray.ysize = nrows
-        data_model.meta.exposure.ngroups = ngroups
-        data_model.meta.exposure.nframes = 3
-
-        # create a saturation model for the saturation step
-        saturation_model = SaturationModel((1032, 1024))
-        saturation_model.meta.instrument.name = "MIRI"
-        saturation_model.meta.instrument.detector = "MIRIMAGE"
-        saturation_model.meta.subarray.xstart = 1
-        saturation_model.meta.subarray.xsize = 1024
-        saturation_model.meta.subarray.ystart = 1
-        saturation_model.meta.subarray.ysize = 1032
-        add_test_refmodel_metadata(saturation_model)
-
-        return data_model, saturation_model
-
-    return _cube
-
-
-@pytest.fixture(scope="function")
-def setup_nrs_irs2_cube():
-    def _cube():
-        # create a JWST datamodel for NIRSPEC IRS2 data
-        data_model = RampModel((1, 5, 3200, 2048))
-        data_model.data = np.ones(((1, 5, 3200, 2048)))
-        data_model.groupdq = np.zeros(((1, 5, 3200, 2048)))
-        data_model.pixeldq = np.zeros(((3200, 2048)))
-        data_model.meta.instrument.name = "NIRSPEC"
-        data_model.meta.instrument.detector = "NRS1"
-        data_model.meta.instrument.filter = "F100LP"
-        data_model.meta.observation.date = "2019-07-19"
-        data_model.meta.observation.time = "23:23:30.912"
-        data_model.meta.exposure.type = "NRS_LAMP"
-        data_model.meta.subarray.name = "FULL"
-        data_model.meta.subarray.xstart = 1
-        data_model.meta.subarray.xsize = 2048
-        data_model.meta.subarray.ystart = 1
-        data_model.meta.subarray.ysize = 2048
-        data_model.meta.exposure.nrs_normal = 16
-        data_model.meta.exposure.nrs_reference = 4
-        data_model.meta.exposure.readpatt = "NRSIRS2"
-        data_model.meta.exposure.nframes = 5
-        data_model.meta.exposure.ngroups = 5
-
-        # create a saturation model for the saturation step
-        saturation_model = SaturationModel((2048, 2048))
-        saturation_model.data = (
-            np.ones((2048, 2048)) * 60000
-        )  # saturation limit for every pixel is 60000
-        saturation_model.meta.subarray.xstart = 1
-        saturation_model.meta.subarray.xsize = 2048
-        saturation_model.meta.subarray.ystart = 1
-        saturation_model.meta.subarray.ysize = 2048
-        add_test_refmodel_metadata(saturation_model)
-
-        # create a bias model for group 2 saturation checking
-        bias_model = SuperBiasModel((3200, 2048))
-        bias_model.data = np.ones((3200, 2048)) * 15000  # bias for every pixel is 15000
-        bias_model.meta.instrument.name = "NIRSPEC"
-        bias_model.meta.instrument.detector = "NRS1"
-        bias_model.meta.subarray.xstart = 1
-        bias_model.meta.subarray.xsize = 2048
-        bias_model.meta.subarray.ystart = 1
-        bias_model.meta.subarray.ysize = 2048
-        bias_model.meta.exposure.readpatt = "NRSIRS2"
-        add_test_refmodel_metadata(bias_model)
-
-        return data_model, saturation_model, bias_model
-
-    return _cube
-
-
-@pytest.fixture(scope="function")
-def setup_nrs_nrs_cube():
-    def _cube():
-        # create a JWST datamodel for NIRSPEC NRS-mode data
-        data_model = RampModel((1, 5, 3200, 2048))
-        data_model.data = np.ones(((1, 5, 3200, 2048)))
-        data_model.groupdq = np.zeros(((1, 5, 3200, 2048)))
-        data_model.pixeldq = np.zeros(((3200, 2048)))
-        data_model.meta.instrument.name = "NIRSPEC"
-        data_model.meta.instrument.detector = "NRS1"
-        data_model.meta.instrument.filter = "F100LP"
-        data_model.meta.observation.date = "2019-07-19"
-        data_model.meta.observation.time = "23:23:30.912"
-        data_model.meta.exposure.type = "NRS_LAMP"
-        data_model.meta.subarray.name = "FULL"
-        data_model.meta.subarray.xstart = 1
-        data_model.meta.subarray.xsize = 2048
-        data_model.meta.subarray.ystart = 1
-        data_model.meta.subarray.ysize = 2048
-        data_model.meta.exposure.nrs_normal = 16
-        data_model.meta.exposure.nrs_reference = 4
-        data_model.meta.exposure.readpatt = "NRS"
-        data_model.meta.exposure.nframes = 4
-        data_model.meta.exposure.ngroups = 5
-
-        # create a saturation model for the saturation step
-        saturation_model = SaturationModel((2048, 2048))
-        saturation_model.data = (
-            np.ones((2048, 2048)) * 60000
-        )  # saturation limit for every pixel is 60000
-        saturation_model.meta.instrument.name = "NIRSPEC"
-        saturation_model.meta.instrument.detector = "NRS1"
-        saturation_model.meta.subarray.xstart = 1
-        saturation_model.meta.subarray.xsize = 2048
-        saturation_model.meta.subarray.ystart = 1
-        saturation_model.meta.subarray.ysize = 2048
-        add_test_refmodel_metadata(saturation_model)
-
-        return data_model, saturation_model
-
-    return _cube
-
-
-@pytest.fixture(scope="function")
-def setup_nis_superstripe_cube():
-    """Set up mock NIRISS superstripe data to test."""
-
-    def _cube():
-        data_model = make_superstripe_model()
-        num_stripes = data_model.meta.subarray.num_superstripe
-        data_model.pixeldq = np.zeros((num_stripes, *data_model.data.shape[-2:]), dtype=np.uint32)
-
-        saturation_model = SaturationModel((2048, 2048))
-        saturation_model.meta.subarray.xstart = 1
-        saturation_model.meta.subarray.ystart = 1
-        saturation_model.meta.subarray.xsize = 2048
-        saturation_model.meta.subarray.ysize = 2048
-        saturation_model.meta.instrument.name = "NIRISS"
-        add_test_refmodel_metadata(saturation_model)
-
-        # create a bias model
-        bias_model = SuperBiasModel((2028, 2048))
-        bias_model.data = np.ones((2048, 2048)) * 15000  # bias for every pixel is 15000
-        bias_model.meta.instrument.name = "NIRISS"
-        bias_model.meta.instrument.detector = "NIS"
-        bias_model.meta.subarray.xstart = 1
-        bias_model.meta.subarray.xsize = 2048
-        bias_model.meta.subarray.ystart = 1793
-        bias_model.meta.subarray.ysize = 256
-        bias_model.meta.subarray.fastaxis = -2
-        bias_model.meta.subarray.slowaxis = -1
-        bias_model.meta.exposure.readpatt = "ANY"
-        add_test_refmodel_metadata(bias_model)
-
-        return data_model, saturation_model, bias_model
-
-    return _cube
