+++ jwst/superbias/tests/test_bias_sub.py
@@ -3,10 +3,12 @@
 import numpy as np
 from stdatamodels.jwst.datamodels import dqflags
 
+from jwst.saturation.tests.helpers import setup_nis_superstripe_cube
 from jwst.superbias.bias_sub import do_correction
+from jwst.superbias.tests.helpers import setup_full_cube, setup_subarray_cube
 
 
-def test_basic_superbias_subtraction(setup_full_cube):
+def test_basic_superbias_subtraction():
     """Check basic superbias subtraction."""
 
     # Create inputs, data, and superbiases
@@ -29,7 +31,7 @@ def test_basic_superbias_subtraction(setup_full_cube):
     assert np.array_equal(output.data, manual)
 
 
-def test_subarray_correction(setup_subarray_cube):
+def test_subarray_correction():
     """Check that the proper subarray is extracted from the full frame
     reference file during subtraction."""
 
@@ -57,7 +59,7 @@ def test_subarray_correction(setup_subarray_cube):
     assert np.array_equal(output.data, manualout)
 
 
-def test_dq_propagation(setup_full_cube):
+def test_dq_propagation():
     """Check that the PIXELDQ array of the science exposure is correctly
     combined with the reference file DQ array."""
 
@@ -81,7 +83,7 @@ def test_dq_propagation(setup_full_cube):
     assert output.pixeldq[5, 5] == dqval1 + dqval2
 
 
-def test_nans_in_superbias(setup_full_cube):
+def test_nans_in_superbias():
     """Check that no superbias subtraction is done for pixels that have a
     value of NaN in the reference file."""
 
@@ -109,7 +111,7 @@ def test_nans_in_superbias(setup_full_cube):
     assert np.array_equal(output.data[0, :, 50, 50], data.data[0, :, 50, 50] - blevel)
 
 
-def test_zeroframe(setup_full_cube):
+def test_zeroframe():
     """ """
     darr1 = [11800.0, 11793.0, 11823.0, 11789.0, 11857.0]
     darr2 = [11800.0, 11793.0, 11823.0, 11789.0, 11857.0]
@@ -144,11 +146,11 @@ def test_zeroframe(setup_full_cube):
     np.testing.assert_equal(check, output.zeroframe[0, 0, :])
 
 
-def test_superstripe_correction(setup_nis_superstripe_cube):
+def test_superstripe_correction():
     """Smoke test for superstripe handling."""
 
     # Create superstripe model and full frame bias model
-    data, bias = setup_nis_superstripe_cube()
+    data, _, bias = setup_nis_superstripe_cube()
 
     # The bias model has a single flat value
     blevel = bias.data[0, 0]
